requirement.txt
Django==2.0
djangorestframework==3.9.0
Faker==0.9.0
pkg-resources==0.0.0
python-dateutil==2.7.4
pytz==2018.6
six==1.11.0
text-unidecode==1.2
